Authorization on API

+++ BlazorApp/Controllers/DoctorController.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Shared.DTO.Core;
 using Services.Core;
+using System.IdentityModel.Tokens.Jwt;
 
 namespace BlazorApp.Controllers;
 
@@ -16,18 +17,30 @@ public class DoctorController : ControllerBase
     }
 
     [HttpGet]
-    public async Task<ActionResult<IEnumerable<DoctorDTO>>> GetDoctors()
+    public async Task<ActionResult<IEnumerable<DoctorDTO>>> GetDoctors([FromHeader] string Authorization)
     {
         try
         {
-            var doctors = await _service.GetDoctors();
-
-            if (doctors == null)
+            if (string.IsNullOrEmpty(Authorization))
             {
-                return NotFound();
+                return Unauthorized("JWT token is missing in Authorization header");
             }
 
-            return Ok(doctors);
+            if (IsAuthorized(Authorization, "Employee"))
+            {
+                var doctors = await _service.GetDoctors();
+
+                if (doctors == null)
+                {
+                    return NotFound();
+                }
+
+                return Ok(doctors);
+            } 
+            else
+            {
+                return Unauthorized("User is not authorized to access this resource");
+            }
         }
         catch (Exception ex)
         {
@@ -37,19 +50,30 @@ public class DoctorController : ControllerBase
     }
 
     [HttpPost]
-    public async Task<ActionResult<DoctorDTO>> CreateDoctor([FromBody] DoctorDTO newDoc)
+    public async Task<ActionResult<DoctorDTO>> CreateDoctor([FromBody] DoctorDTO newDoc, [FromHeader] string Authorization)
     {
         try
         {
-            var doctor = await _service.CreateDoctor(newDoc);
-
-            if (doctor == null)
+            if (string.IsNullOrEmpty(Authorization))
             {
-                return BadRequest();
+                return Unauthorized("JWT token is missing in Authorization header");
             }
 
-            return Ok(doctor);
+            if (IsAuthorized(Authorization, "Admin"))
+            {
+                var doctor = await _service.CreateDoctor(newDoc);
 
+                if (doctor == null)
+                {
+                    return BadRequest();
+                }
+
+                return Ok(doctor);
+            }
+            else
+            {
+                return Unauthorized("User is not authorized to access this resource");
+            }
         }
         catch (Exception ex)
         {
@@ -59,19 +83,30 @@ public class DoctorController : ControllerBase
     }
 
     [HttpPut]
-    public async Task<ActionResult<DoctorDTO>> UpdateDoctor([FromBody] DoctorDTO newDoc)
+    public async Task<ActionResult<DoctorDTO>> UpdateDoctor([FromBody] DoctorDTO newDoc, [FromHeader] string Authorization)
     {
         try
         {
-            var doctor = await _service.UpdateDoctor(newDoc);
-
-            if (doctor == null)
+            if (string.IsNullOrEmpty(Authorization))
             {
-                return BadRequest();
+                return Unauthorized("JWT token is missing in Authorization header");
             }
 
-            return Ok(doctor);
+            if (IsAuthorized(Authorization, "Admin"))
+            {
+                var doctor = await _service.UpdateDoctor(newDoc);
 
+                if (doctor == null)
+                {
+                    return BadRequest();
+                }
+
+                return Ok(doctor);
+            }
+            else
+            {
+                return Unauthorized("User is not authorized to access this resource");
+            }
         }
         catch (Exception ex)
         {
@@ -81,17 +116,55 @@ public class DoctorController : ControllerBase
     }
 
     [HttpDelete]
-    public async Task<ActionResult> DeleteDoctor([FromBody] DoctorDTO docToDelete)
+    public async Task<ActionResult> DeleteDoctor([FromBody] DoctorDTO docToDelete, [FromHeader] string Authorization)
     {
         try
         {
-            await _service.DeleteDoctor(docToDelete);
-            return Ok();
+            if (string.IsNullOrEmpty(Authorization))
+            {
+                return Unauthorized("JWT token is missing in Authorization header");
+            }
+
+            if (IsAuthorized(Authorization, "Admin"))
+            {
+                await _service.DeleteDoctor(docToDelete);
+                return Ok();
+            }
+            else
+            {
+                return Unauthorized("User is not authorized to access this resource");
+            }
         }
         catch (Exception ex)
         {
             Console.WriteLine(ex);
             return StatusCode(500, "Internal server error");
         }
+    }
+
+    private static bool IsAuthorized(string jwtToken, string checkRole)
+    {
+        try
+        {
+            var handler = new JwtSecurityTokenHandler();
+            var jsonToken = handler.ReadToken(jwtToken) as JwtSecurityToken;
+
+            if (jsonToken == null)
+            {
+                return false;
+            }
+
+            var roleClaims = jsonToken.Claims.Where(c => c.Type == "http://schemas.microsoft.com/ws/2008/06/identity/claims/role")?.Select(c => c.Value);
+            var concatenatedRoles = string.Join(",", roleClaims);
+            if (roleClaims != null && concatenatedRoles.Contains(checkRole, StringComparison.OrdinalIgnoreCase))
+            {
+                return true;
+            }
+        }
+        catch (Exception ex)
+        {
+            return false;
+        }
+        return false;
     }
 }
